@@ -28,18 +28,19 @@
     "lint": "nx workspace-lint && ng lint",
     "e2e": "nx run transloco-playground-e2e:e2e",
     "update": "nx migrate latest",
+    "migrate": "nx migrate --run-migrations",
     "dep-graph": "nx dep-graph",
     "pre-commit": "node scripts/pre-commit.js && lint-staged"
   },
   "dependencies": {
-    "@angular/animations": "12.2.5",
-    "@angular/common": "12.2.5",
-    "@angular/compiler": "12.2.5",
-    "@angular/core": "12.2.5",
-    "@angular/forms": "12.2.5",
-    "@angular/platform-browser": "12.2.5",
-    "@angular/platform-browser-dynamic": "12.2.5",
-    "@angular/router": "12.2.5",
+    "@angular/animations": "13.3.0",
+    "@angular/common": "13.3.0",
+    "@angular/compiler": "13.3.0",
+    "@angular/core": "13.3.0",
+    "@angular/forms": "13.3.0",
+    "@angular/platform-browser": "13.3.0",
+    "@angular/platform-browser-dynamic": "13.3.0",
+    "@angular/router": "13.3.0",
     "@messageformat/core": "3.0.0",
     "bootstrap": "5.1.0",
     "chalk": "4.1.2",
@@ -83,35 +84,37 @@
     "@types/jasmine": "3.8.2",
     "@types/jest": "27.0.2",
     "@types/node": "16.7.2",
-    "@typescript-eslint/eslint-plugin": "4.33.0",
-    "@typescript-eslint/parser": "4.33.0",
+    "@typescript-eslint/eslint-plugin": "5.10.2",
+    "@typescript-eslint/parser": "5.10.2",
     "cypress": "8.7.0",
     "dotenv": "10.0.0",
-    "eslint": "7.32.0",
+    "eslint": "8.7.0",
     "eslint-config-prettier": "8.3.0",
     "eslint-plugin-cypress": "2.11.3",
     "git-cz": "4.7.1",
     "husky": "7.0.2",
-    "jasmine-core": "3.9.0",
+    "jasmine-core": "3.10.1",
     "jasmine-spec-reporter": "7.0.0",
     "jest": "27.2.3",
-    "jest-preset-angular": "10.0.1",
-    "karma": "6.3.4",
+    "jest-preset-angular": "11.1.1",
+    "karma": "6.3.17",
     "karma-chrome-launcher": "3.1.0",
+    "karma-coverage": "~2.2.0",
     "karma-coverage-istanbul-reporter": "3.0.3",
     "karma-jasmine": "4.0.1",
     "karma-jasmine-html-reporter": "1.7.0",
     "lint-staged": "11.1.2",
-    "ng-packagr": "12.2.4",
+    "ng-packagr": "13.2.1",
+    "nx": "13.9.3",
     "postcss": "8.3.11",
     "postcss-import": "14.0.2",
     "postcss-preset-env": "6.7.0",
     "postcss-url": "10.1.3",
-    "prettier": "2.3.2",
+    "prettier": "2.6.0",
     "ts-jest": "27.0.5",
     "ts-mocks": "3.0.1",
     "ts-node": "10.2.1",
-    "typescript": "4.3.5"
+    "typescript": "4.5.5"
   },
   "config": {
     "commitizen": {
